test/GradeBook.Tests/BookTests.cs
using System;
using Xunit;


namespace GradeBook.Tests
{
    public class BookTests
    {
        [Fact]
        public void Test1()
        {
            //Given
            var x = GetInt();
            SetInt(ref x);
            //When

            //Then
            Assert.Equal(42, x);
        }

        private void SetInt(ref int x)
        {
            x = 42;
        }

        private int GetInt()
        {
            return 3;
        }

        [Fact]
        public void BookCalculatesStats()
        {
            // arrange portion -- staging the data and variables you want to test
            var book = new Book("");
            book.AddGrade(89.1);
            book.AddGrade(90.5);
            book.AddGrade(77.3);

            // act portion -- store the calculations or procedures you want to test
            var result = book.GetStats();

            // assert portion -- assert something about the value that was computed
            Assert.Equal(85.6, result.Average, 1);
            Assert.Equal(90.5, result.High, 1);
            Assert.Equal(77.3, result.Low, 1);
            Assert.Equal('B', result.Letter);
        }
    }
}
